+++ data_struct/unrolled_linked_list/unrolled_linked_list.h
@@ -8,8 +8,8 @@
 
 /* Forward Declaration
  */
-template <typename T, size_t NodeSize> class Unrolled_List;
-template <typename T, size_t NodeSize>
+template <typename T, int NodeSize> class Unrolled_List;
+template <typename T, int NodeSize>
 std::ostream& operator<<(std::ostream &os, const Unrolled_List<T, NodeSize> &rhs);
 
 
@@ -22,12 +22,12 @@ std::ostream& operator<<(std::ostream &os, const Unrolled_List<T, NodeSize> &rhs
  * than 1 element. This is done to improve cahce perfromance, but increases overhead in memory and
  * inserting.
  */
-template <typename T, size_t NodeSize = 8>
+template <typename T, int NodeSize = 8>
 class Unrolled_List
 {
     private: // Variables
         List_Node<T, NodeSize> *head, *tail;
-        size_t N;
+        int N;
 
     private: // Methods
         void split_node(List_Node<T, NodeSize> *ptr);
@@ -36,7 +36,7 @@ class Unrolled_List
         // Constructors and destructors.
         Unrolled_List();
         template <class InputIt> Unrolled_List(InputIt begin, InputIt end);
-        Unrolled_List(size_t cnt);
+        Unrolled_List(int cnt);
         Unrolled_List(std::initializer_list<T> lst);
         Unrolled_List(const Unrolled_List &rhs);
         Unrolled_List(Unrolled_List &&rhs);
@@ -44,7 +44,7 @@ class Unrolled_List
 
         // Capacity
         bool empty() const noexcept;
-        size_t size() const noexcept;
+        int size() const noexcept;
 
         // Access
         T& back();
@@ -54,8 +54,8 @@ class Unrolled_List
 
         // Modifiers
         void clear() noexcept;
-        void insert(size_t pos, const T &item);
-        void insert(size_t pos, T &&item);
+        void insert(int pos, const T &item);
+        void insert(int pos, T &&item);
         void push_back(const T &item);
         void push_back(T &&item);
         void push_front(const T &item);
